moving the healthchecks to a Extensions issue #84
