+++ src/Bet.AspNetCore.HealthChecks/Bet.AspNetCore.HealthChecks.csproj
@@ -24,6 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Bet.Extensions.HealthChecks\Bet.Extensions.HealthChecks.csproj" />
     <ProjectReference Include="..\Bet.Extensions\Bet.Extensions.csproj" />
   </ItemGroup>
