+++ src/Bet.AspNetCore.HealthChecks/HealthChecksApplicationBuilderExtensions.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
+
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
