+++ src/Bet.Extensions.HealthChecks/AzureBlobStorage/AzureBlobStorageHealthCheck.cs
@@ -11,7 +11,7 @@ using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Bet.AspNetCore.HealthChecks.AzureBlobStorage
+namespace Bet.Extensions.HealthChecks.AzureBlobStorage
 {
     public class AzureBlobStorageHealthCheck : IHealthCheck
     {
