+++ src/Bet.Extensions.HealthChecks/Bet.Extensions.HealthChecks.csproj
@@ -4,8 +4,21 @@
     <TargetFrameworks>netstandard2.0;netstandard2.1</TargetFrameworks>
   </PropertyGroup>
 
+  <PropertyGroup>
+    <Description>The collection of the HealthChecks functionality for Worker Pods.</Description>
+    <PackageTags>dotnetcore, worker, healthcheck, hosting, extensions</PackageTags>
+  </PropertyGroup>
+
+
   <ItemGroup>
+    <PackageReference Include="Microsoft.Azure.Services.AppAuthentication" />
+    <PackageReference Include="Microsoft.Azure.Storage.Blob" />
     <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" />
+    <PackageReference Include="Microsoft.Extensions.Http.Polly" />
+    <PackageReference Include="Newtonsoft.Json" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Bet.Extensions\Bet.Extensions.csproj" />
+  </ItemGroup>
 </Project>
